moved the formatting to CommandNode
